added filedialog and track from coeffs

+++ simulering.py
@@ -1,6 +1,10 @@
 import math
 import numpy as np
 from matplotlib import pyplot as plt
+from iptrack import *
+
+import tkinter as tk
+from tkinter import filedialog
 
 ## CONSTANTS
 g = 9.81  # m/s
@@ -55,15 +59,9 @@ def normal_g_force(v_tangential, krumning):
     return 1 + v_tangential**2 * krumning / g
 
 
-def bane(x):
-    return 3433.2281869052126*x**15 + -35425.80967983638*x**14 + \
-            164086.51268783736*x**13 + -449939.19577019487*x**12 + \
-            810832.0563168445*x**11 + -1007456.1154303564*x**10 + \
-            880102.071263024*x**9 + -540502.2906810011*x**8 + \
-            228926.39127614404*x**7 + -63888.210540241824*x**6 + \
-            10617.08175653279*x**5 + -759.7060900777494*x**4 + \
-            -33.33911458989251*x**3 + 8.555075814607894*x**2 + \
-            -1.7645318855075836*x**1 + 0.6532652470202632
+def bane(x, coeffs):
+    # len - 1 because the last term is a constant, not x**1, and i starts at 0
+    return sum(c * x**(len(coeffs) - 1 - i) for i, c in enumerate(coeffs))
 
 
 def write_csv(*args, filename, header, ext="csv"):
@@ -84,6 +82,15 @@ def write_csv(*args, filename, header, ext="csv"):
 
 
 def main():
+    print("Enter file with data from Tracker")
+    root = tk.Tk()
+    root.withdraw()
+    filename = filedialog.askopenfilename()
+    root.destroy()  # removes window instance
+
+    coeffs = iptrack(filename)
+    func = lambda x: bane(x, coeffs)
+
     step_h = 0.001  # s, eulers method step
     UPPER_T = 1.1  # s, how long to simulate
 
@@ -99,35 +106,38 @@ def main():
     dfdx = [0.0] * len(time_)  # first derivative
     d2fdx2 = [0.0] * len(time_)  # second derivative
 
-    angle[0] = alpha(bane, pos_x[0])
-    pos_y[0] = bane(pos_x[0])
+    angle[0] = alpha(func, pos_x[0])
+    pos_y[0] = func(pos_x[0])
 
     # calc points for curve
     t = [i * step_h for i in range(0, int(UPPER_T / step_h))]
     b = list()
     for x in t:
-        b.append(bane(x))
+        b.append(func(x))
 
     # simulation loop
     for n in time_[:-1]:  # n from 0 to len-1
-        angle[n+1] = alpha(bane, pos_x[n])
+        angle[n+1] = alpha(func, pos_x[n])
         accel_tangential[n+1] = accel_at_x(angle[n])
         speed_tangential[n+1] = speed_tangential[n] + step_h * accel_tangential[n]
         pos_x[n+1] = pos_x[n] + step_h * x_comp(speed_tangential[n], angle[n])
         pos_y[n+1] = pos_y[n] - step_h * y_comp(speed_tangential[n], angle[n])
 
-        curvature[n+1] = krumning(bane, pos_x[n])
-        normal[n+1] = normal_g_force(speed_tangential[n], krumning(bane, pos_x[n]))
+        curvature[n+1] = krumning(func, pos_x[n])
+        normal[n+1] = normal_g_force(speed_tangential[n], krumning(func, pos_x[n]))
 
-        dfdx[n] = first_derivative(bane, pos_x[n])
-        d2fdx2[n] = second_derivative(bane, pos_x[n])
+        dfdx[n] = first_derivative(func, pos_x[n])
+        d2fdx2[n] = second_derivative(func, pos_x[n])
 
 
-    # write simulation to file
-    write_csv(time_, angle, pos_x, pos_y, speed_tangential, accel_tangential,\
-            normal, curvature, dfdx, d2fdx2, filename="krumbane1_sim",
-            header=["time","angle","x","y","speed_tangential",\
-            "accel_tangential","normal_force","curvature","dfdx","d2fdx2"])
+
+    ans = input("Do you want to write simulation results to file? (y/N) ").lower()
+    if ans.startswith("y"):
+        # write simulation to file
+        write_csv(time_, angle, pos_x, pos_y, speed_tangential, accel_tangential,\
+                normal, curvature, dfdx, d2fdx2, filename="krumbane1_sim",
+                header=["time","angle","x","y","speed_tangential",\
+                "accel_tangential","normal_force","curvature","dfdx","d2fdx2"])
 
     # plotting
     print("Plotting...")
